backend/requirements.txt
numpy==1.22.3
pandas==1.4.2
pydantic==1.10.13
fastapi==0.85.0
uvicorn==0.16.0
PyMySQL==1.0.2
multimethod==1.8
python-multipart==0.0.5
PyMySQL==1.0.2
SQLAlchemy==1.4.37
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0
fastapi-pagination==0.10.0
pillow
requests
cryptography
opencv-python
APScheduler==3.9.1
